Move NovelAI endpoints to separate file
